# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "接口文档",
        "title": "框架接口",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.0.1"
    },
    "host": "192.168.2.89:8082",
    "paths": {
        "/test": {
            "get": {
                "description": "Test Swagger",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "测试接口",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "userId",
                        "name": "some_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/email/reset": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "重置邮箱，根据邮件中验证码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "重置邮箱",
                "parameters": [
                    {
                        "description": "email,validate_code",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.resetEmailParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success data为null",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "token错误 data为null",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "无法访问 data为null",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "内部错误 data为null",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/user/email/reset/validate": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "发送重置邮箱验证码邮件",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "发送重置邮箱验证码邮件",
                "parameters": [
                    {
                        "description": "email",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.emailParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success data为null",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "token错误 data为null",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "无法访问 data为null",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "内部错误 data为null",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/user/info": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "编辑用户信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "编辑用户信息",
                "parameters": [
                    {
                        "description": "sex:0男，1女，2保密 age:年龄 user_name:用户名，昵称 name:实际姓名 info:其他信息",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.userInfoParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "调用成功 data为null",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "token错误 data为null",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "无法访问 data为null",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "内部错误 data为null",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "description": "登录接口，通过email，手机号或用户名",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "登录",
                "parameters": [
                    {
                        "description": "log_id:email，手机号或用户名 password_encrypt:加密后的密码",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.loginParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success data为{\\\"token\\\":token, \\\"user_id\\\":user id}",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "token错误 data为null",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "无法访问 data为null",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "内部错误 data为null",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/user/password/reset": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "重置密码（已登录状态）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "重置密码（已登录状态）",
                "parameters": [
                    {
                        "description": "密码（加密过的）",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.passwordEncryptParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success data为null",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "token错误 data为null",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "无法访问 data为null",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "内部错误 data为null",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/user/password/reset/email": {
            "post": {
                "description": "重置密码（发送邮件方式），根据邮件中验证码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "重置密码（发送邮件方式）",
                "parameters": [
                    {
                        "description": "email",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.registerOrResetPasswordByEmailParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success data为null",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "token错误 data为null",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "无法访问 data为null",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "内部错误 data为null",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/user/password/reset/email/validate": {
            "post": {
                "description": "邮件发送重置密码验证码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "邮件发送重置密码验证码",
                "parameters": [
                    {
                        "description": "邮箱",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.emailParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success data为null",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "token错误 data为null",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "无法访问 data为null",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "内部错误 data为null",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/user/register/email": {
            "post": {
                "description": "邮件注册",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "邮件注册",
                "parameters": [
                    {
                        "description": "邮箱",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.registerOrResetPasswordByEmailParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success data为{\\\"id\": user id}",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "token错误 data为null",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "无法访问 data为null",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "内部错误 data为null",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/user/register/email/validate": {
            "post": {
                "description": "邮件发送注册验证码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "邮件发送注册验证码",
                "parameters": [
                    {
                        "description": "邮箱",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.emailParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success data为null",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "token错误 data为null",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "无法访问 data为null",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "内部错误 data为null",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "service.emailParams": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                }
            }
        },
        "service.loginParams": {
            "type": "object",
            "properties": {
                "log_id": {
                    "type": "string"
                },
                "password_encrypt": {
                    "type": "string"
                }
            }
        },
        "service.passwordEncryptParams": {
            "type": "object",
            "properties": {
                "password_encrypt": {
                    "type": "string"
                }
            }
        },
        "service.registerOrResetPasswordByEmailParams": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password_encrypt": {
                    "type": "string"
                },
                "validate_code": {
                    "type": "string"
                }
            }
        },
        "service.resetEmailParams": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "validate_code": {
                    "type": "string"
                }
            }
        },
        "service.userInfoParams": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "string"
                },
                "info": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "sex": {
                    "type": "string"
                },
                "user_name": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}